Rename _loader->core
