+++ r2c_isg/loaders/__init__.py
@@ -1,1 +1,1 @@
-from ._loader import Loader
+from .core import Loader
